+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - `Union::is_subset_of` [MR106](https://github.com/Qrlew/qrlew/pull/106)
+- Replaced `Expr::filter_column_data_type` by `DataType::filter`[MR104](https://github.com/Qrlew/qrlew/pull/104)
+
 
 ## [0.2.3] - 2023-09-04
 ### Changed